+++ my_exercises/testing/functions.py
@@ -1,19 +1,5 @@
 from typing import Union
 
 
-def divide(dividend: Union[int, float], divisor: Union[int, float]):
-    if divisor == 0:
-        raise ValueError("The divisor cannot be zero.")
-    return dividend / divisor
-
-
-def multiply(*args: Union[int, float]):
-    print(args)
-    if len(args) == 0:
-        raise ValueError('At least one value to multiply must be passed')
-
-    total = 1
-    for arg in args:
-        total *= arg
-
-    return total
+def divide(dividend: Union[int, float], divisior: Union[int, float]):
+    return dividend / divisior
